refactor: extract layer rules into analyzer/layers.rs (#190)
Moves LayerViolation, check_layer_rules, and the AuditResult inherent
methods filter_by_layers + apply_layer_weights into a dedicated module.
The methods are still inherent impls on AuditResult — Rust allows
splitting inherent impls across files in the same crate.
